+++ debug-tools-server/src/main/java/io/github/future0923/debug/tools/server/http/handler/AllClassLoaderHttpHandler.java
@@ -1,15 +1,13 @@
 package io.github.future0923.debug.tools.server.http.handler;
 
 import com.sun.net.httpserver.Headers;
+import io.github.future0923.debug.tools.base.classloader.DefaultClassLoader;
 import io.github.future0923.debug.tools.common.protocal.http.AllClassLoaderRes;
-import io.github.future0923.debug.tools.common.utils.DebugToolsJvmUtils;
 import io.github.future0923.debug.tools.server.DebugToolsBootstrap;
 import org.codehaus.groovy.reflection.SunClassLoader;
 
 import java.lang.instrument.Instrumentation;
-import java.util.ArrayList;
 import java.util.HashSet;
-import java.util.List;
 import java.util.Map;
 import java.util.Set;
 import java.util.concurrent.ConcurrentHashMap;
@@ -23,41 +21,24 @@ public class AllClassLoaderHttpHandler extends BaseHttpHandler<Void, AllClassLoa
 
     public static final String PATH = "/allClassLoader";
 
-    public static ClassLoader defaultClassLoader;
-
     public static final Map<String, ClassLoader> classLoaderMap = new ConcurrentHashMap<>();
-
-    public static final List<String> baseClassList = new ArrayList<>();
-
-    static {
-        baseClassList.add("org.slf4j.Logger");
-        baseClassList.add("org.apache.log4j.Logger");
-        baseClassList.add("org.springframework.beans.factory.BeanFactory");
-    }
 
     private AllClassLoaderHttpHandler() {
 
     }
 
+    public static ClassLoader getDebugToolsClassLoader() {
+        return AllClassLoaderHttpHandler.class.getClassLoader();
+    }
+
     @Override
     protected AllClassLoaderRes doHandle(Void req, Headers responseHeaders) {
         AllClassLoaderRes res = new AllClassLoaderRes();
-        String mainClass = DebugToolsJvmUtils.getMainClass();
+        res.setDefaultIdentity(Integer.toHexString(System.identityHashCode(DefaultClassLoader.getDefaultClassLoader())));
         Instrumentation instrumentation = DebugToolsBootstrap.INSTANCE.getInstrumentation();
         Set<AllClassLoaderRes.Item> allClassLoaderResSet = new HashSet<>();
         for (Class<?> clazz : instrumentation.getAllLoadedClasses()) {
             ClassLoader classLoader = clazz.getClassLoader();
-            if (res.getDefaultIdentity() == null) {
-                if (mainClass != null) {
-                    if (mainClass.equals(clazz.getName())) {
-                        res.setDefaultIdentity(Integer.toHexString(System.identityHashCode(classLoader)));
-                        defaultClassLoader = classLoader;
-                    }
-                } else if (baseClassList.contains(clazz.getName())) {
-                    res.setDefaultIdentity(Integer.toHexString(System.identityHashCode(classLoader)));
-                }
-
-            }
             if (classLoader != null
                     // groovy的加载器不要
                     && !(classLoader instanceof SunClassLoader)