+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+# [3.1.0](https://github.com/future0923/debug-tools/compare/v3.0.1...v3.1.0) (2024-12-17)
+
+- Search http url to jump directly to the corresponding method definition.
+
+# [3.0.1](https://github.com/future0923/debug-tools/compare/v3.0.0...v3.0.1) (2024-12-12)
+
+- Support intellij idea 2024.3 version.
+- Fixed bug where the default ClassLoader could not be obtained when remotely executing Groovy.
+
 # 3.0.0 (2024-11-01)
 
 ### Features